+++ app/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+from flask_login import LoginManager
+
+app = Flask(__name__)
+login_manager = LoginManager(app)
+
+from app import routes
